Initial Commit
At this point, the design should work. I've routed the 1.1V rail that was missing from the initial prototype and fixed the pads for the LEDs. There's still work to be done, and I haven't ordered this iteration of the design yet to verify, but I believe it should work.