Rev 3
JLCPCB noticed a mistake in my order, I had remnants from an early version in the gerbers folder. Cleared that out and made some updates to the board and resubmitted as Rev3. I added a GND via that was missing near a 3.3V via underneath the RP2040, added Rev 3 to the silkscreen, and nudged C207 away from the USB lines and added a GND via. Gerbers, drill, and centroid files regenerated. Rev 3 renders generated as well.